mybatis auto generator

+++ src/main/java/com/khpp/weixin/db/service/UserService.java
@@ -3,7 +3,7 @@ package com.khpp.weixin.db.service;
 import com.khpp.weixin.core.GenericService;
 import com.khpp.weixin.db.domain.User;
 
-public interface UserService extends GenericService<User, Long> {
+public interface UserService extends GenericService<User, Integer> {
 
 	/**
 	 * 用户验证
